Add files via upload
This is a program which is for research, searching where is yeast strain samples stocked. In addition to this, we need two text files and one csv files. In two text files, one file is described yeast sample name you want to search and another file is result of searching. The csv file have information of approximately 4000 samples position name. With this program, we can easily find where is samples you want to use.